\name{@PKGNAME@}
\docType{package}

\alias{@PKGNAME@-package}
\alias{@PKGNAME@}

\title{@PKGTITLE@}

\description{
  This annotation package stores genomic scores.
}

\note{
  This data package was made from resources at @PROVIDER@
  based on the version @PROVIDERVERSION@.
}

\author{@AUTHOR@}

\seealso{
  \code{\link[GenomicScores]{GenomicScores}}
}

\examples{
library(@PKGNAME@)

ls("package:@PKGNAME@")
@PKGNAME@
}

\keyword{package}
\keyword{data}
